Converte chave de licenças para inteiro
A função que busca as licenças dos servidores estava buscando os CMs
como string e com zeros à esquerda. Isso causava problemas na hora de
comparar com a base de dados, que tem os CMs como inteiros.

Para resolver isso, converti a chave das licenças para inteiro.

+++ src/importador_excel.py
@@ -16,7 +16,7 @@ def obtem_tempos_licencas() -> dict[int, int]:
         sql_query = f.read()
     df_licencas = BancoDeDados().realiza_consulta(sql_query)
     # Cria o dicionário indexado por cm
-    return dict(zip(df_licencas["cm"], df_licencas["qtde_dias_licenca"]))
+    return dict(zip(df_licencas["cm"].astype(int), df_licencas["qtde_dias_licenca"]))
 
 
 class ImportadorProjecaoExcel:
